+++ server/auth/auth_test.go
@@ -2,6 +2,7 @@ package auth
 
 import (
 	"testing"
+	"time"
 
 	"github.com/remind101/empire"
 	"github.com/stretchr/testify/assert"
@@ -12,34 +13,35 @@ func TestStaticAuthenticator(t *testing.T) {
 	u := &empire.User{}
 	a := StaticAuthenticator("username", "password", "otp", u)
 
-	user, err := a.Authenticate("badusername", "password", "otp")
+	session, err := a.Authenticate("badusername", "password", "otp")
 	assert.Equal(t, ErrForbidden, err)
-	assert.Nil(t, user)
+	assert.Nil(t, session)
 
-	user, err = a.Authenticate("username", "badpassword", "otp")
+	session, err = a.Authenticate("username", "badpassword", "otp")
 	assert.Equal(t, ErrForbidden, err)
-	assert.Nil(t, user)
+	assert.Nil(t, session)
 
-	user, err = a.Authenticate("username", "password", "badotp")
+	session, err = a.Authenticate("username", "password", "badotp")
 	assert.Equal(t, ErrTwoFactor, err)
-	assert.Nil(t, user)
+	assert.Nil(t, session)
 
-	user, err = a.Authenticate("username", "password", "otp")
+	session, err = a.Authenticate("username", "password", "otp")
 	assert.NoError(t, err)
-	assert.Equal(t, u, user)
+	assert.Equal(t, NewSession(u), session)
 }
 
 func TestMultiAuthenticator_First(t *testing.T) {
 	u := &empire.User{}
+	s := NewSession(u)
 	a1 := new(mockAuthenticator)
 	a2 := new(mockAuthenticator)
 	a := MultiAuthenticator(a1, a2)
 
-	a1.On("Authenticate", "username", "password", "").Return(&empire.User{}, nil)
+	a1.On("Authenticate", "username", "password", "").Return(s, nil)
 
-	user, err := a.Authenticate("username", "password", "")
+	session, err := a.Authenticate("username", "password", "")
 	assert.NoError(t, err)
-	assert.Equal(t, u, user)
+	assert.Equal(t, s, session)
 
 	a1.AssertExpectations(t)
 	a2.AssertExpectations(t)
@@ -47,16 +49,17 @@ func TestMultiAuthenticator_First(t *testing.T) {
 
 func TestMultiAuthenticator_Second(t *testing.T) {
 	u := &empire.User{}
+	s := NewSession(u)
 	a1 := new(mockAuthenticator)
 	a2 := new(mockAuthenticator)
 	a := MultiAuthenticator(a1, a2)
 
 	a1.On("Authenticate", "username", "password", "").Return(nil, ErrForbidden)
-	a2.On("Authenticate", "username", "password", "").Return(&empire.User{}, nil)
+	a2.On("Authenticate", "username", "password", "").Return(s, nil)
 
-	user, err := a.Authenticate("username", "password", "")
+	session, err := a.Authenticate("username", "password", "")
 	assert.NoError(t, err)
-	assert.Equal(t, u, user)
+	assert.Equal(t, s, session)
 
 	a1.AssertExpectations(t)
 	a2.AssertExpectations(t)
@@ -70,9 +73,9 @@ func TestMultiAuthenticator_None(t *testing.T) {
 	a1.On("Authenticate", "username", "password", "").Return(nil, ErrForbidden)
 	a2.On("Authenticate", "username", "password", "").Return(nil, ErrForbidden)
 
-	user, err := a.Authenticate("username", "password", "")
+	session, err := a.Authenticate("username", "password", "")
 	assert.Equal(t, ErrForbidden, err)
-	assert.Nil(t, user)
+	assert.Nil(t, session)
 
 	a1.AssertExpectations(t)
 	a2.AssertExpectations(t)
@@ -85,23 +88,71 @@ func TestMultiAuthenticator_ErrTwoFactor(t *testing.T) {
 
 	a1.On("Authenticate", "username", "password", "").Return(nil, ErrTwoFactor)
 
-	user, err := a.Authenticate("username", "password", "")
+	session, err := a.Authenticate("username", "password", "")
 	assert.Equal(t, ErrTwoFactor, err)
-	assert.Nil(t, user)
+	assert.Nil(t, session)
 
 	a1.AssertExpectations(t)
 	a2.AssertExpectations(t)
+}
+
+func TestWithMaxSessionDuration(t *testing.T) {
+	t1 := time.Date(2009, time.November, 10, 23, 0, 0, 0, time.UTC)
+	m := new(mockAuthenticator)
+	a := WithMaxSessionDuration(m, func() time.Time { return t1 })
+
+	m.On("Authenticate", "username", "password", "").Return(&Session{}, nil)
+
+	session, err := a.Authenticate("username", "password", "")
+	assert.NoError(t, err)
+	assert.Equal(t, t1, *session.ExpiresAt)
+
+	m.AssertExpectations(t)
+}
+
+func TestWithMaxSessionDuration_WithShorterExpiresAt(t *testing.T) {
+	s1 := time.Date(2009, time.November, 10, 23, 0, 0, 0, time.UTC)
+	m := new(mockAuthenticator)
+	a := WithMaxSessionDuration(m, func() time.Time { return s1.Add(24 * time.Hour) })
+
+	exp := s1.Add(12 * time.Hour)
+	m.On("Authenticate", "username", "password", "").Return(&Session{
+		ExpiresAt: &exp,
+	}, nil)
+
+	session, err := a.Authenticate("username", "password", "")
+	assert.NoError(t, err)
+	assert.Equal(t, exp, *session.ExpiresAt)
+
+	m.AssertExpectations(t)
+}
+
+func TestWithMaxSessionDuration_WithLongerExpiresAt(t *testing.T) {
+	s1 := time.Date(2009, time.November, 10, 23, 0, 0, 0, time.UTC)
+	m := new(mockAuthenticator)
+	a := WithMaxSessionDuration(m, func() time.Time { return s1.Add(12 * time.Hour) })
+
+	exp := s1.Add(24 * time.Hour)
+	m.On("Authenticate", "username", "password", "").Return(&Session{
+		ExpiresAt: &exp,
+	}, nil)
+
+	session, err := a.Authenticate("username", "password", "")
+	assert.NoError(t, err)
+	assert.Equal(t, s1.Add(12*time.Hour), *session.ExpiresAt)
+
+	m.AssertExpectations(t)
 }
 
 type mockAuthenticator struct {
 	mock.Mock
 }
 
-func (m *mockAuthenticator) Authenticate(username, password, otp string) (*empire.User, error) {
+func (m *mockAuthenticator) Authenticate(username, password, otp string) (*Session, error) {
 	args := m.Called(username, password, otp)
-	user := args.Get(0)
-	if user != nil {
-		return user.(*empire.User), args.Error(1)
+	session := args.Get(0)
+	if session != nil {
+		return session.(*Session), args.Error(1)
 	}
 	return nil, args.Error(1)
 }